pyproject.toml
[tool.poetry]
name = "stepflow"
version = "0.1.0"
description = ""
authors = ["Zhihao Liu <liuzhihao0206@gmail.com>"]
readme = "README.md"

packages = [
    { include = "stepflow" }
]

[tool.poetry.dependencies]
python = "^3.13"
alembic = "^1.15.2"
sqlalchemy = "^2.0.40"
fastapi = "^0.115.12"
pytest = "^8.3.5"
aiosqlite = "^0.21.0"
pytest-asyncio = "^0.26.0"
pytz = "^2025.2"
greenlet = "^3.1.1"
aiohttp = "^3.11.14"
uvicorn = "^0.34.0"
httpx = "^0.28.1"
requests = "^2.32.3"
websockets = "^15.0.1"
pyyaml = "^6.0.2"
jsonschema = "^4.23.0"
pydantic = "^2.11.4"
jsonpath-ng = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
